Collect load functions
